--- management-app/src/components/BoardColumn/BoardColumn.types.ts
@@ -1,3 +1,0 @@
-import { TColumn } from '../../api/types';
-
-export type BoardColumnProps = TColumn;
