+++ management-app/src/pages/BoardPage/components/BoardColumn/BoardColumn.types.ts
@@ -1,0 +1,3 @@
+import { TColumn } from '../../../../api/types';
+
+export type BoardColumnProps = TColumn;
