+++ management-app/src/pages/BoardPage/components/BoardColumn/index.ts
