+++ management-app/src/pages/BoardPage/components/ColumnTitle/ColumnTitle.types.ts
