+++ management-app/src/pages/BoardPage/components/TaskCard/TaskCard.types.ts
@@ -1,4 +1,4 @@
-import { TTaskResponse } from '../../api/types';
+import { TTaskResponse } from '../../../../api/types';
 
 export type TaskCardProps = {
   taskInfo: TTaskResponse;
