+++ management-app/src/pages/BoardPage/index.ts
@@ -1,0 +1,1 @@
+export * from './BoardPage';
